Added special events for RETURN and ESC key presses.
The RETURN-key special event responder is defaultKey, the
ESC-key special event responder is escKey. It can be defined
on the active component or any of its parents. The state
(enabled/disabled) is ignored and the event traverses down
to the app via the parent links. If any responder returns true,
the chain is stopped. For the first level, the siblings
are traversed, but after that, only parents.