Changed component inheritance structure and values
The component hierarchy of components is now:
  HClass -> HValueResponder -> HView -> HControl

This means HView can handle values, and it does so via HValueResponder. This also means the remaining difference between HView and HControl is that HControl handles UI events, but HView does not.

Due to the changes in class hierarchy, there's now a #setOptions method to extend for options overrides. HView and HControl also supports everything from a single options object, which should include rect and parent object references. The general idea in the future is to store most of the instance properties in the options object directly and provide getters and setters with validation on the HView or HControl class. This will allow suspending component states in HValue values some time in the future.